Split out event-tracking file into separated files for  functions
