+++ src/event-tracking/find-trackable.js
@@ -1,0 +1,11 @@
+'use strict'
+
+const buildListener = require('./build-listener')
+
+module.exports = () => {
+  const elementsToTrack = Array.prototype.slice.call(document.querySelectorAll('[data-click-events]'))
+
+  if (elementsToTrack && typeof ga === 'function') {
+    buildListener(elementsToTrack)
+  }
+}
