--- test/helpers/setup-browser-env.js
@@ -1,11 +1,0 @@
-const ga = (send, type, category, action, label) => {
-  return {
-    send,
-    type,
-    category,
-    action,
-    label
-  }
-}
-
-require('browser-env')({window: {ga}})
